FHE now works! (for secure parameter setting)
So the problem was the treatment of the plaintext. Let "base" be the base we work on. After a multiplication the value of the plaintext in each coefficient of the polynomial grows. Once it goes past T, it cannot be retrieved correctly.

So how much it grows? Say the plaintext has a value in the top most coefficient. After a multiplication that one grows by roughly 2d times! So   [base *(2d)]^mults <= T  (and this is a rough bound)

Either way, for your setting of the parameters you need base = 2 (and it was 8!) and mults = 1.